requirements.txt
appdirs==1.4.3
asgiref==3.2.10
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
distro==1.4.0
Django==3.0.8
django-cors-headers==3.4.0
django-import-export==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jdcal==1.4.1
lockfile==0.12.2
MarkupPy==1.14
msgpack==0.6.2
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.8
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
